+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - ContextMenu builder
+- TreeTableView column builder that takes an observable value creator
 
 ## [1.4.8] - 2016-05-20